Create separate ES suggest module
