+++ stroom-search/stroom-search-elastic/src/main/java/stroom/search/elastic/suggest/ElasticSuggestConfig.java
@@ -1,0 +1,33 @@
+package stroom.search.elastic.suggest;
+
+import stroom.util.config.annotations.RequiresRestart;
+import stroom.util.config.annotations.RequiresRestart.RestartScope;
+import stroom.util.shared.AbstractConfig;
+import stroom.util.shared.IsStroomConfig;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import com.fasterxml.jackson.annotation.JsonPropertyDescription;
+import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+
+@JsonPropertyOrder(alphabetic = true)
+public class ElasticSuggestConfig extends AbstractConfig implements IsStroomConfig {
+
+    private final Boolean enabled;
+
+    public ElasticSuggestConfig() {
+        enabled = true;
+    }
+
+    @SuppressWarnings("unused")
+    @JsonCreator
+    public ElasticSuggestConfig(@JsonProperty("enabled") final Boolean enabled) {
+        this.enabled = enabled;
+    }
+
+    @JsonPropertyDescription("Suggest terms in query expressions when using an Elasticsearch index as the data source.")
+    @RequiresRestart(RestartScope.UI)
+    public Boolean getEnabled() {
+        return enabled;
+    }
+}
