+++ stroom-search/stroom-search-elastic/src/main/java/stroom/search/elastic/suggest/ElasticSuggestionsQueryHandler.java
@@ -1,4 +1,4 @@
-package stroom.search.elastic;
+package stroom.search.elastic.suggest;
 
 import stroom.core.query.SuggestionsQueryHandler;
